requirements.txt
# ── Core UI ──────────────────────────────────────────────────────────────────
streamlit>=1.32.0

# ── Groq LLM API ─────────────────────────────────────────────────────────────
groq>=0.9.0

# ── Local embeddings (no API key needed) ─────────────────────────────────────
sentence-transformers>=3.0.0

# ── Vector search ─────────────────────────────────────────────────────────────
# faiss-cpu is optional but recommended for speed.
# The app falls back to pure-NumPy cosine similarity if it isn't available.
faiss-cpu>=1.7.4

# ── Document parsing ──────────────────────────────────────────────────────────
PyMuPDF>=1.23.0          # PDF text + image extraction
python-docx>=1.1.0       # Word (.docx) text + image extraction

# ── Numerics / Image handling ─────────────────────────────────────────────────
numpy>=1.24.0
Pillow>=10.0.0

# ── Environment variables (local dev) ─────────────────────────────────────────
python-dotenv>=1.0.0
